fixed clippy pedantic warns

+++ src/model.rs
@@ -132,6 +132,6 @@ impl EmployeeCategory {
     #[allow(clippy::cast_precision_loss)]
     pub fn cost_per_millisecond(&self) -> f64 {
         // 2000 work hours per year, 60 mins per hours, 60 secs per hour, 1000 ms per second.
-        self.salary / (2000.0 * 60.0 * 60.0 * 1000.0)
+        (self.salary / (2000 * 60 * 60 * 1000)) as f64
     }
 }
